requirements.txt
# Core ML/AI Libraries
openai>=2.21.0              # OpenAI API client
instructor>=1.14.0          # Structured LLM outputs with Pydantic
pydantic>=2.12.0            # Data validation using type hints
pandas>=3.0.0               # Data manipulation and analysis
numpy>=2.0.0                # Numerical computing (installed with pandas)

# Visualization
seaborn>=0.13.0             # Statistical data visualization
matplotlib>=3.10.0          # Plotting library (installed with seaborn)

# Data / Benchmarking
datasets>=4.0.0             # HuggingFace datasets for benchmark loading

# Utilities
python-dotenv>=1.2.0        # Environment variable management

# API
fastapi>=0.115.0            # Web framework
uvicorn[standard]>=0.30.0   # ASGI server
httpx>=0.27.0               # HTTP client

# Testing
pytest>=7.0.0               # Testing framework
pytest-cov>=4.0.0           # Coverage reporting